requirements.txt
langchain==0.1.12
langchain-google-genai==0.0.11
gpt4all==2.0.2
fastapi==0.110.0
Appium-Python-Client==3.2.1
#Following packages needed for langserve
langserve==0.0.51
orjson==3.9.15
sse-starlette==1.8.2
uvicorn==0.28.0
click==8.1.7
httpx-sse==0.4.0 